requirements.txt
center-tk-window==1.0.0
mysql-connector==2.2.9
pika==1.1.0
pillow==8.0.1
pystray==0.17.1
python-dotenv==0.14.0
pywin32==228
requests==2.24.0
future~=0.18.2
setproctitle~=1.1.10